+++ src/value/mod.rs
@@ -11,11 +11,10 @@ mod unitset;
 
 pub use self::colors::{Color, Hsla, Hwba, Rgba};
 pub use self::list_separator::ListSeparator;
-pub use self::number::Number;
+pub use self::number::{Number, Rational};
 pub use self::numeric::Numeric;
 pub use self::operator::Operator;
 pub use self::quotes::Quotes;
 pub use self::unit::{Dimension, Unit};
 pub use self::unitset::UnitSet;
-pub use num_rational::Rational;
 pub(crate) use range::{RangeError, ValueRange};